+++ packages/plugin-link/web/src/toolbar/TextLinkPanel.jsx
@@ -1,5 +1,6 @@
 import { Component } from 'react';
 import PropTypes from 'prop-types';
+import { EditorState } from 'draft-js';
 import { isEmpty } from 'lodash';
 import {
   insertLinkAtCurrentSelection,
@@ -26,7 +27,8 @@ export default class TextLinkPanel extends Component {
       t,
       isActive: !isEmpty(linkData),
       onDone: this.createLinkEntity,
-      onCancel: this.hideLinkPanel,
+      onCancel: this.onCancel,
+      hidePanel: this.hideLinkPanel,
       onDelete: this.deleteLink,
       onOverrideContent: this.props.onOverrideContent,
       uiSettings,
@@ -60,6 +62,18 @@ export default class TextLinkPanel extends Component {
     const editorState = getEditorState();
     const newEditorState = removeLinksInSelection(editorState, setEditorState);
     setEditorState(newEditorState);
+  };
+
+  onCancel = () => {
+    const { getEditorState, setEditorState } = this.props;
+    const editorState = getEditorState();
+    const selection = editorState.getSelection();
+    const newEditorState = EditorState.forceSelection(
+      editorState,
+      selection.merge({ anchorOffset: selection.focusOffset })
+    );
+    setEditorState(newEditorState);
+    this.hideLinkPanel();
   };
 
   hideLinkPanel = () => {